Create an ansible role for Jenkins on Windows (adapted from geerlingguy.ansible-role-jenkins)
Started from geerlingguy.ansible-role-jenkins role and modify it to make it work on a windows target.
The initial goal was to make the role compatible with both windows and linux but this is (I found) adding too many duplicate tasks (due to the fact windows and linux modules are never common)
This role uses chocolatey as a package manager for windows and uses powershell commands (instead of linux shell).
I added 2 groovy scripts: one to set the proxy and one to set the JLNP port (without this the CLI is disabled)
I also changed the way to bypass the security wizard.

Still some work to do on the configuration file (wich is totaly diferent than config files under linux) to set things like the jenkins port.